pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "dsv-wrapper"
version = "0.1.0"
description = "Reusable Python package for accessing DSV systems (Daisy, Handledning)"
readme = "README.md"
requires-python = ">=3.12"
authors = [
    {name = "Edwin", email = "edwin@dsv.su.se"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
dependencies = [
    "pydantic>=2.10.0",
    "requests>=2.32.0",
    "aiohttp>=3.11.0",
    "beautifulsoup4>=4.12.0",
    "python-dotenv>=1.0.0",
    "lxml>=5.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "black>=24.0.0",
    "ruff>=0.7.0",
]

[project.urls]
Homepage = "https://github.com/edwin/dsv-wrapper"
Repository = "https://github.com/edwin/dsv-wrapper"

[tool.setuptools.packages.find]
where = ["."]
include = ["dsv_wrapper*"]

[tool.black]
line-length = 100
target-version = ["py312"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP"]
ignore = []
